feat: move LogClick, LogScreen components
- components 폴더 내로 분리했습니다.
