+++ src/demo/App.tsx
@@ -1,7 +1,7 @@
-import { BrowserRouter } from 'react-router-dom';
-import { Routes, Route } from 'react-router-dom';
+import { Routes, Route, BrowserRouter } from 'react-router-dom';
 import { Home } from './Home';
 import { Drawer } from './Drawer';
+
 export const App = () => {
   return (
     <>
